JAMES-1797 don't dispose MimeMessageInputStreamSource file too early
